Update patch set 2
Patch Set 2:

Andreas: This change was done to formalize our team's definitions of done into a document that we can refer to later. We chose to add it in CONTRIBUTING.md, and since it's a public document on GitHub we realized it would be confusing to only include definitions of done with no context on how to contribute for external contributors.

Would like you to review this, as it would also send a message that we are open for outside PR contributions.

Patch-set: 2
Label: Code-Review=0