Update patch set 1
Patch Set 1:

Hans-Peter, this is a changes for the next Debian release Buster which will be available in ~6 month... so we are trying to create Vagrant VM and check what we have inside. There will be separate VM for you to test now new code version works for us before releasing and making any decissiong.

There should be no changes in the current and nearest releases.

Patch-set: 1