Update patch set 4
Patch Set 4: Code-Review-1

Declined for now, as Oliver wants to stick with the Sipwise-green instead of the pastel tone. This change is going to be revived once we receive the final color scheme which is supposed to be migrated to a complete "dark-mode" theme anyways.

Patch-set: 4
Reviewer: Andreas Granig <1000007@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=-1