TT#129212 Optimise the initial load of the user
TAC:
Parallelise requests in the user load function which do not depend on each other

Note: it can be optimized better if we lazyload RTCengine and initialize it in a lazy way. But it should be implemented as a separate ticket.

(cherry picked from commit 5a3093e837ca702791e3c2de93c56c7de1497e9d)

Change-Id: I35601f447c3f19c0a720c3848bf69c8d53d29395