TT#138660 Fix to build release mr6.5.13
The jobs fail with:

  https://jenkins.mgm.sipwise.com/job/ngcp-csc-ui-binaries/3580/console

  ...
  21:46:51 npm ERR! fetch failed https://npm-registry.sipwise.com/define-properties/-/define-properties-1.1.2.tgz
  21:46:51 npm WARN retry will retry, error on last attempt: Error: certificate has expired
  21:46:51 npm ERR! fetch failed https://npm-registry.sipwise.com/clap/-/clap-1.2.3.tgz
  21:46:51 npm WARN retry will retry, error on last attempt: Error: certificate has expired
  21:46:51 npm ERR! fetch failed https://npm-registry.sipwise.com/circular-json/-/circular-json-0.3.3.tgz
  21:46:51 npm WARN retry will retry, error on last attempt: Error: certificate has expired
  ...

This is due to the known problem with Let's Encrypt expiration of
certificates as of late.  Ideally the npm-registry.sipwise.com
certificate would be re-generated with the new chain so as to avoid this
problem, but it is not possible right now in short notice.

Build-depending on ca-certificates to include the necessary updates and
get the problem fixed did not help, maybe it needs updates to crypto
libraries or something else (there were known problems in some of them
by which they didn't accept the certificate if a single of the chains
was invalid).

Resorting to using HTTP instead of HTTPS also didn't work, the server
returns 404 for the same URLs.

So trying to disable strict SSL chekcing at the npm level.

mr6.5.13 is the last release of mr6.5 and the whole mr6.5* is going to
become immeditely obsolete, so even if not ideal it should be OK to have
as a workaround as it's almost at the finish line -- almost everyone is
using newer releases by now.

Change-Id: I712197d5a33d7a31a54958bc2b0290b8f9d58e78