TT#134104 Issue - password change error shown Russian message ignoring choosen language
NOTE: please AR carefully as the change affects every CSC request

Additions:
After testing the approach of rewriting the url property of the request object, problems appeared. The library vue-resource considers the url property actually as url path and not as the entire url. This is clearly a misleading naming in vue-resource. But for query params it offers the params object.

Change-Id: I27ca475df9ae8cde487907a0e341984e9c384241