MT#58743 Create a new Pbx Configuration -> Customer Preferences page
With the creation of the Customer Preferences Page we add the possibility for the subscriber's administrator to view and edit some of the preferences that are applied customer's wide.

We also refactored $errMsg, the function that handles validation error messages globally.

The reviewed function displays the error text based on the errorMessages methods in src/validators/index.js

Change-Id: Ic9a1b2eab9b71255026447a690d3426097cbaf93