MT#60150 sip username changes in pbx group creation
sip username is created removing whitespaces and special chars from
the name input. This logic was handled by a lodash method. This commit
replaces the mentioned method with a custom function that mirrors the
logic we use in ngcp-admin to handle the pbx group creation.
The new method is less aggressive in fact it allows the use of
underscores and doesn't dash numbers. As part of the change we also
expose sip username as read-only field in the edit pbxGroupDetails
page.

Change-Id: I086b3a7011e47d8f542ac92f42bd923c38630331