MT#60583 Amend Login API calls and 403 Invalid Licence err
We have introduced the concept of licenses which can be active
or inactive (all active licenses are stored in the array "licenses"
-GET 'api/platforminfo'). This means that a service (e.g. fax)
can be activated, but the license to use it could be active or
inactive. If the latter is true, the user is not allowed to
use the service in fact the backend returns a 403 - Invalid
license when the relevant endpoints are called.
* Before this commit we were retrieving the fax services as
part of the login phase, but with the new licenses logic the
existing flow was broken by a 403 for those users with inactive
fax license and it was impossible to login into the platform.
With this commit we amend getUserData() to wrap the call to
retrieve the fax settings in a statement that checks if the
fax licence is active.
* We edit the "Invalid license" message returned by the backend
with a more user-friendly text. The text is showed in a
small banner on top of the page every time the user opens
a page that is not supposed to access because the relevant
license is inactive.
* We remove some unused code in CscLayoutMain.vue
* v-if moved before class to silence warning in AuiMobileAppBadges

Change-Id: Ie7831e0024475c3a5aa8acf3874e6eda442fc9a2
(cherry picked from commit 98eee9c1ff2579d16709384bd56da73f86726ece)