MT#60545 Block outg/incom calls refactoring
We were using some shared logic to set subscriber preferences
for incoming and outcoming calls preferences. However the
PATCH/'replace' request was always failing with a 422 before
running a second block of code to finally add the new
preference. This commit adds a custom block of logic to
handle the creation of subscriber preferences for the
outg/incom calls.

Change-Id: I42a4b82a1b805e75597c1037caf4f789e2b287b6