MT#60711 removing lang parameter when we use api v2
Due to some changes on api v2, we cannot pass a parameter that doesn't exist to api.
In csc, we put the lang parameter in each call but it returns an error when using api v2.
So in the case we use api v2 we just removed this parameter.

Change-Id: Ib9e5decc67178d65626a285688e41b88b16e0b03
(cherry picked from commit 51bc01af9687d124af4d38789787ff1aefe47901)