MT#60689 Add change password page for CSC
The backend returns "403 - Password Expired" when user's
password is too old and we redirect the user to the
/changepassword page where they can update their pw.
Users will be required to insert username, current password,
new password and to retype the new one.
Success: user redirected to login page.
Failure: a banner with the error is shown for 10 seconds.
NOTE:
* Password requirements are always presented in the frontend
  as guidelines but validations are not enforced.
  It was a joint decision with the backend team not to expose
  whether the web_validation is active on the platform.
* If the web_validation is active, the backend will respond
  with a 422 and the error message will describe the missed
  password requirements.

Change-Id: Ic8b34980ff3099712f72042294409a2eb8e9347f
(cherry picked from commit 46357637cdedd88a2ce59b27537fd68596906727)