MT#61314 Change call recordings DELETE
The backend has been amended so that when a call
is recorded on both sides (caller and callee),
if one of the two sides deletes the recording,
the item is only removed from their view and the other
party can still see the recording. To achieve this we
add a param subscriber_id into the delete request.
As well this commit fixes the pop-up appearing to confirm
the user wants to proceed with deletion. The pop-up
was appearing without title and message, now both
are shown.

Change-Id: I7808fd3fad819e3e194448d16e882502f0a29bc6