MT#61843 Create call forward bugfix
Our tests have caught a weird behaviour for the endpoint
/cfdestinationsets v1 (mr11.5). The endpoint returns
an empty body after creating a new destination set.
This has been fixed by the backend team for versions
mr12.5 onward. To fix the problem in mr11.5 we add
an extra GET request to retrieve the new destination
set, before patching the mapping with the updated info.
NGCP-Flow: mr11.5

Change-Id: I3b42ce6682d4b51adea17e6d6bf7a0dd3cbb7860