MT#61843 remove subscriberId from GET destinations request after creating new mapping
When a PBX admin creates a CF for a Seat,
the endpoint (mr11.5x) automatically assigns
the PBX admin id to the destination even
if we pass the desired subscriberId in the
payload. For this reason this commit changes
the GET request params to name only.

NGCP-Flow: mr11.5

Change-Id: I1e2d3520a1d5441997dd519bb5daed29d13e3362
(cherry picked from commit 67465275a79f73fb82ebd704d34476a47e763062)