MT#62062 Reload CF after amending source set Bugfix
When a PBX Seat user tries to delete a sorce set that
was previously added by a PBX admin the backend
return an error as the entity doesn't belong to the
Seat user. In this scenario the source set is not
delete from the DB and it is removed from the relevant
mapping. The missing try/catch didn't allow the
code to reload the updated resources to show the
changes.
NGCP-Flow: mr13.0

Change-Id: Ic02750e4fbc9a47b6cd424ab9331c9fcb7b15799
(cherry picked from commit 621da19cf2eb977e2e1ad2615ec32483cbe28417)