MT#62062 Reload CF after amending source set Bugfix
When a PBX Seat user tries to delete a sorce set that
was previously added by a PBX admin the backend
return an error as the entity doesn't belong to the
Seat user. In this scenario the source set is not
delete from the DB and it is removed from the relevant
mapping. The missing try/catch didn't allow the
code to reload the updated resources to show the
changes.
NGCP-Flow: mr12.5

Change-Id: I2862d5922aa1df99d11b2d6d58dcd8e161160253