MT#62756 CSC now sends Referred-By with callee local data in blind xfer
CSC-UI has been changed so that, during a blind transfer, the
Referred-By header is customized with local identity URI.
Without this customisation, jssip puts blindly the subscriber username
which is not always what kamailio proxy expects.

Having correct Referred-By header is important because kamailio proxy
uses this header in order to identify the transferor in a blind transfer call.
In particular, kamailio proxy checks if Referred-By matches custom header
P-Xfer-Firstcall-Callee-To, in this case B is recogized as transferor.

Change-Id: If149b34c16a5dc13116f8fea9a6d8046498753cd