MT#62760 After Ring timeout CFSeat bugfix
If a pilot admin subscriber tries to change
the After Ring timeout for CFU set on one
of the seats, the CF config crashes
and change is not applied. This happens
because we are not passing the
subscriber_id of the Seat. In fact the
change in the CF is applied to the logged
in user instead. In this commit we pass
the subscriber_id attribute down the three
and improve the error handling.

Change-Id: I27b34a095263d18ca8d00b98bc3a528b4ee35f42