MT#62769 The sendFax icon appears in the PBX Configuration even if the functionality is not enabled
The Send Fax icon was displayed in the PBX Configuration
even when the functionality was not enabled in the fax settings.

This was caused by the ability to activate
the feature at the seat or group level.

The logic has been corrected, and the Send Fax functionality
now appears only when it is properly enabled for the user.

Change-Id: I02358f12eaf1a5543d2c6552e21d4452c56acdc6