MT#61427 Remove access to the OLD Customer Details page
In case the CSC was configured to work in mixed mode
(www_admin.http_csc.csc_js_enable == 'mixed') then the
OLD CSC page of the customer details was displayed on
the new CSC for administrative subscribers.

This was added in commit 03546154289e33e6a93ab4d3d

The OLD CSC is not fully deprecated so also the access
to this page should be removed.

Change-Id: Ic791fb85c2039ce2c3e801b3ca826bee5a4c7e7c