MT#63292 Remove stylus
Remove stylus library as it has a malware
and it is not used in the repo.
We are not backporting the trunk commit
as currently the two branches have huge
differences with regards of libs versions.

NGCP-Flow:mr12.5

Change-Id: I9820133e59cf01a3c55a446ec94d1e55c8c1ed49
(cherry picked from commit 37b0c949207f48a297791a1c053540774012fd0c)