MT#63292 Remove stylus
Remove stylus library as it has a malware
and it is not used in the repo.
We are not backporting the trunk commit
as currently the two branches have huge
differences with regards of libs versions.

NGCP-Flow: mr11.5
Change-Id: Iad196ce9d02e80a5775a1e4f3138e554b1a36d37
(cherry picked from commit a47b394b26a12f0ae80922a5b452989766945cea)