MT#63384 Fix Group Details page crash upon refresh
Fixed an issue where the component tried to access the `id` property
of `this.groupSelected` when it was undefined. This happened because
the state gets cleared on refresh, and the component didn't have a
check to reload the necessary resources.

- Added check to ensure resources are reloaded if state is cleared
- Prevent accessing `id` of undefined in the component

Change-Id: Ib6b669df87d255cf174254b32fa1c451d0901f73
(cherry picked from commit 4e72e67697a6ff162be49bb001eaf54cf8ba4d28)